update metadata reading, should work for most versions of il2cpp now.

+++ UnityDumper/src/IL2CPP/Il2Cpp.h
@@ -7,6 +7,6 @@ std::string GetGameAssemblyPathManual();
 bool GetMetadataPath(const std::string& gamePath, std::string& metadataPath);
 std::string GetMetadataPathManual();
 
-bool IsMetadataSupported(int32_t version);
+bool IsMetadataSupported(float version);
 
 void AnalyzeIl2Cpp(std::string gamePath);
